Refactor to make internal services private
