+++ test/controllers/ReferenceDataControllerSpec.scala
@@ -35,14 +35,12 @@ import play.api.test.Helpers._
 import services.ingestion.ReferenceDataService.DataProcessingResult.DataProcessingFailed
 import services.ingestion.ReferenceDataService.DataProcessingResult.DataProcessingSuccessful
 import services.ingestion.ReferenceDataService
-import services.ingestion.SchemaValidationService
 
 import scala.concurrent.Future
 
 class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with BeforeAndAfterEach {
 
-  val mockReferenceDataService    = mock[ReferenceDataService]
-  val mockSchemaValidationService = mock[SchemaValidationService]
+  val mockReferenceDataService = mock[ReferenceDataService]
 
   private val testJson = Json.obj("foo" -> "bar")
 
@@ -118,7 +116,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
 
   override def beforeEach(): Unit = {
     super.beforeEach()
-    Mockito.reset(mockReferenceDataService, mockSchemaValidationService)
+    Mockito.reset(mockReferenceDataService)
   }
 
   // Do not use directly use `app` instead
@@ -126,8 +124,7 @@ class ReferenceDataControllerSpec extends SpecBase with GuiceOneAppPerSuite with
     new GuiceApplicationBuilder()
       .configure("play.http.router" -> "testOnlyDoNotUseInAppConf.Routes")
       .overrides(
-        bind[ReferenceDataService].toInstance(mockReferenceDataService),
-        bind[SchemaValidationService].toInstance(mockSchemaValidationService)
+        bind[ReferenceDataService].toInstance(mockReferenceDataService)
       )
       .build()
 }
