+++ test/services/ListRetrievalServiceSpec.scala
@@ -16,8 +16,6 @@
 
 package services
 
-import java.time.LocalDate
-
 import base.SpecBase
 import generators.BaseGenerators
 import generators.ModelArbitraryInstances
